reformat list filters, rm Header
