+++ src/library/PoolList/Default.tsx
@@ -16,7 +16,11 @@ import { useTheme } from 'contexts/Themes';
 import { useUi } from 'contexts/UI';
 import { Tabs } from 'library/Filter/Tabs';
 import { usePoolFilters } from 'library/Hooks/usePoolFilters';
-import { Header, List, Wrapper as ListWrapper } from 'library/List';
+import {
+  FilterHeaderWrapper,
+  List,
+  Wrapper as ListWrapper,
+} from 'library/List';
 import { MotionContainer } from 'library/List/MotionContainer';
 import { Pagination } from 'library/List/Pagination';
 import { SearchInput } from 'library/List/SearchInput';
@@ -32,7 +36,6 @@ export const PoolList = ({
   disableThrottle,
   allowSearch,
   pools,
-  title,
   defaultFilters,
   allowListFormat = true,
 }: PoolListProps) => {
@@ -178,27 +181,6 @@ export const PoolList = ({
 
   return (
     <ListWrapper>
-      <Header>
-        <div>
-          <h4>{title}</h4>
-        </div>
-        {allowListFormat && (
-          <div>
-            <button type="button" onClick={() => setListFormat('row')}>
-              <FontAwesomeIcon
-                icon={faBars}
-                color={listFormat === 'row' ? colors.primary[mode] : 'inherit'}
-              />
-            </button>
-            <button type="button" onClick={() => setListFormat('col')}>
-              <FontAwesomeIcon
-                icon={faGripVertical}
-                color={listFormat === 'col' ? colors.primary[mode] : 'inherit'}
-              />
-            </button>
-          </div>
-        )}
-      </Header>
       <List $flexBasisLarge={allowMoreCols ? '33.33%' : '50%'}>
         {allowSearch && poolsDefault.length > 0 && (
           <SearchInput
@@ -206,31 +188,58 @@ export const PoolList = ({
             placeholder={t('search')}
           />
         )}
-        <Tabs
-          config={[
-            {
-              label: t('all'),
-              includes: [],
-              excludes: [],
-            },
-            {
-              label: t('active'),
-              includes: ['active'],
-              excludes: ['locked', 'destroying'],
-            },
-            {
-              label: t('locked'),
-              includes: ['locked'],
-              excludes: [],
-            },
-            {
-              label: t('destroying'),
-              includes: ['destroying'],
-              excludes: [],
-            },
-          ]}
-          activeIndex={1}
-        />
+        <FilterHeaderWrapper>
+          <div>
+            <Tabs
+              config={[
+                {
+                  label: t('all'),
+                  includes: [],
+                  excludes: [],
+                },
+                {
+                  label: t('active'),
+                  includes: ['active'],
+                  excludes: ['locked', 'destroying'],
+                },
+                {
+                  label: t('locked'),
+                  includes: ['locked'],
+                  excludes: [],
+                },
+                {
+                  label: t('destroying'),
+                  includes: ['destroying'],
+                  excludes: [],
+                },
+              ]}
+              activeIndex={1}
+            />
+          </div>
+          <div>
+            {allowListFormat && (
+              <div>
+                <button type="button" onClick={() => setListFormat('row')}>
+                  <FontAwesomeIcon
+                    icon={faBars}
+                    color={
+                      listFormat === 'row' ? colors.primary[mode] : 'inherit'
+                    }
+                  />
+                </button>
+                <button type="button" onClick={() => setListFormat('col')}>
+                  <FontAwesomeIcon
+                    icon={faGripVertical}
+                    color={
+                      listFormat === 'col' ? colors.primary[mode] : 'inherit'
+                    }
+                  />
+                </button>
+              </div>
+            )}
+          </div>
+        </FilterHeaderWrapper>
+
         {pagination && poolsToDisplay.length > 0 && (
           <Pagination page={page} total={totalPages} setter={setPage} />
         )}
